requirements.txt
# =============================================================================
# ROCKET-2 Python Dependencies
# =============================================================================
# macOS ARM64 (Apple Silicon) 安装步骤：
#
#   1. pip install torch torchvision
#   2. pip install minestudio==1.1.2 --no-deps
#   3. pip install -r requirements.txt
#   4. cd MineStudio/minestudio/utils/realtime_sam && pip install --no-build-isolation -e .
#   5. python -m minestudio.simulator.entry -y
#
# Linux GPU 服务器安装步骤：
#
#   1. pip install torch torchvision
#   2. pip install minestudio==1.1.2
#   3. pip install -r requirements.txt
#   4. cd MineStudio/minestudio/utils/realtime_sam && pip install --no-build-isolation -e .
#   5. python -m minestudio.simulator.entry -y
#
# =============================================================================

# --- Vision ---
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
timm>=1.0.0
pillow==11.0.0

# --- Web UI ---
gradio>=5.12,<6

# --- Model Utilities ---
einops>=0.8.0
rich>=13.0.0
huggingface_hub>=0.20.0
safetensors>=0.4.0

# --- MineStudio runtime deps (installed via --no-deps on macOS) ---
gym>=0.26.0
gymnasium>=1.0.0
gym3>=0.3.3
av>=13.0.0
lmdb>=1.4.0
hydra-core>=1.3.2
hydra-colorlog>=1.2.0
Pyro4>=4.82
x-transformers>=1.30.0

# --- Training (optional) ---
lightning>=2.2.0
wandb>=0.16.0

# --- Misc ---
requests>=2.31.0
pyyaml>=6.0
numpy<2
